requirements.txt
asgiref==3.5.1
autopep8==1.6.0
certifi==2020.6.20
coverage==6.3.2
Django==3.2.13
django-cors-headers==3.11.0
django-environ==0.8.1
django-resized==0.3.11
django-simpleui==3.9.1
djangorestframework==3.13.1
mssql-django==1.1.3
Pillow==9.1.0
pycodestyle==2.8.0
pyodbc==4.0.32
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
wincertstore==0.2
